High Precision Scheduler + Exception bug fix
The new scheduler runs with really good precision, but needs more testing.
Fixed a bug when a function may throw an exception, which may not be handled and the program killed... except we don't have an operating system to execute _kill ang get_pid, so we get an linker error instead... the error does not mention nothing about exceptions, only _kill and get_pid couldn't be found. While disassembling a debug elf, i've noticed the .section with the exception. Fixed by adding -fno-exceptions.